requirements.txt
ultralytics
opencv-python
numpy
tqdm
pyyaml
pandas
matplotlib
onnx
onnxruntime-gpu
python-dotenv
